requirements.txt
flask==2.3.3
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
markdown==3.4.4
beautifulsoup4==4.12.2
numpy==1.24.3
requests==2.31.0
torch==2.0.1
transformers==4.30.2
colorama>=0.4.6
rich>=13.4.2
pytest>=7.3.1
aiohttp>=3.8.4 